requirements.txt
aniso8601==3.0.2
asn1crypto==0.24.0
astroid==2.2.5
authorizenet==1.1.3
beautifulsoup4==4.7.1
blessings==1.6.1
bpython==0.17
cachetools==3.1.0
certifi==2017.11.5
cffi==1.11.4
chardet==3.0.4
click==6.6
colorama==0.4.1
configobj==5.0.6
coverage==3.6
cryptography==2.1.4
curtsies==0.2.11
Django==2.2.1
django-ckeditor==5.6.1
django-cors-headers==2.5.2
django-crispy-forms==1.7.2
django-debug-toolbar==1.11
django-filter==2.1.0
django-js-asset==1.2.2
django-jsonfield==1.1.0
django-mathfilters==0.4.0
django-microservices==0.1
django-notify-x==0.1.9
django-phone-field==1.7.1
django-redis==4.10.0
django-tastypie==0.14.2
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
drf-writable-nested==0.5.1
Faker==0.8.9
Flask==0.11.1
google-api-python-client==1.7.8
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.3.0
graphene==2.1.3
graphene-django==2.2.0
graphql-core==2.1
graphql-relay==0.4.5
greenlet==0.4.12
httplib2==0.12.1
idna==2.6
isort==4.3.15
itsdangerous==0.24
Jinja2==2.8
lazy-object-proxy==1.3.1
lxml==4.3.3
MarkupSafe==0.23
mccabe==0.6.1
numpy==1.16.2
oauthlib==3.0.1
opencv-python==4.0.0.21
Pillow==5.3.0
promise==2.2.1
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycparser==2.18
Pygments==2.2.0
PyJWT==1.7.1
pylint==2.3.1
python-dateutil==2.6.1
python-mimeparse==1.6.0
pytz==2017.3
PyXB==1.2.5
redis==3.2.1
requests==2.18.4
requests-oauthlib==1.2.0
rsa==4.0
Rx==1.6.1
selenium==3.141.0
simplejson==3.8.2
singledispatch==3.4.0.3
six==1.11.0
soupsieve==1.9.1
sqlparse==0.3.0
text-unidecode==1.1
tidalapi==0.5.0
typed-ast==1.3.1
ua-parser==0.3.6
uritemplate==3.0.0
urllib3==1.22
wcwidth==0.1.7
Werkzeug==0.15.2
wrapt==1.11.1
